Fixed small bug with AxeExtension and ShovelExtension

+++ src/main/java/com/temporal/api/core/registry/factory/extension/item/AxeExtension.java
@@ -11,11 +11,11 @@ import java.util.function.Supplier;
 
 @SuppressWarnings("unchecked")
 public interface AxeExtension {
-    default RegistryObject<AxeItem> createAxe(String name, Tier tier, int damage, float speed) {
+    default RegistryObject<AxeItem> createAxe(String name, Tier tier, float damage, float speed) {
         return this.createAxe(name, tier, damage, speed, new Item.Properties());
     }
 
-    default RegistryObject<AxeItem> createAxe(String name, Tier tier, int damage, float speed, Item.Properties properties) {
+    default RegistryObject<AxeItem> createAxe(String name, Tier tier, float damage, float speed, Item.Properties properties) {
         ItemFactory itemFactory = InjectionContext.getInstance().getObject(ItemFactory.class);
         return (RegistryObject<AxeItem>) itemFactory.createTyped(name, () -> new AxeItem(tier, damage, speed, properties));
     }
